Remove outdated pot files
These files are outdated and not needed, translation is done using our
translation server. The current scripts work even when these files do
not exist. Once there are translated files, the files will be imported
automatically again and updated.

This is part of implementing the new infra setup for translations,
see spec
http://specs.openstack.org/openstack-infra/infra-specs/specs/translation_setup.html
for full details.

Remove test for valid po/pot files from tox.ini, this is now done in the
infra scripts.

Change-Id: I49dded30597533434a7c751cdabf8b6a11e1c1c5